create TodoDescription component
